+++ lib/src/utils/code-splitting/checkAsyncAndDownload.ts
@@ -1,5 +1,3 @@
-import { RouteComponent } from '../types'
-
 export async function checkAsyncAndDownload(component: RouteComponent) {
   const isAsync = /(\.then)/i.test(component.toString())
   if (!isAsync) return component